+++ .php-cs-fixer.php
@@ -1,16 +1,27 @@
-<?php declare(strict_types=1);
+<?php 
+
+declare(strict_types=1);
+
 /*
- * Copyright (c) Cristiano Cinotti 2021.
- *
- * This file is part of siad-pdf-compressor package, release under the APACHE-2 license.
+ * Copyright (c) Cristiano Cinotti
+ * 
+ * This file is part of susina/config-builder package, released under the APACHE-2 license.
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
 
-$config = new Susina\CodingStandard\Config();
-$config->getFinder()
-    ->in(__DIR__ . '/src')
-    ->in(__DIR__ . '/tests')
-;
+$finder = new PhpCsFixer\Finder()->in(__DIR__ . '/src')->in(__DIR__ . '/tests');
 
-return $config;
+return new PhpCsFixer\Config()
+    ->setRules([
+        '@PER-CS3x0' => true,
+        'header_comment' => [
+            'header' => "Copyright (c) Cristiano Cinotti
+
+This file is part of susina/config-builder package, released under the APACHE-2 license.
+For the full copyright and license information, please view the LICENSE file distributed 
+with this source code.",
+        ]
+    ])
+    ->setFinder($finder)
+;