+++ src/Loader/XmlFileLoader.php
@@ -1,9 +1,13 @@
-<?php declare(strict_types=1);
+<?php
+
+declare(strict_types=1);
+
 /*
- * Apache-2 License.
- * This file is part of susina/config-builder package, release under the APACHE-2 license.
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
+ * Copyright (c) Cristiano Cinotti
+ *
+ * This file is part of susina/config-builder package, released under the APACHE-2 license.
+ * For the full copyright and license information, please view the LICENSE file distributed
+ * with this source code.
  */
 
 namespace Susina\ConfigBuilder\Loader;
@@ -21,11 +25,15 @@ use Symfony\Component\Config\Loader\FileLoader;
  */
 class XmlFileLoader extends FileLoader
 {
-    private bool $keepFirstTag = false;
+    /**
+     * The xml converter object.
+     * @var Converter
+     */
+    private Converter $converter;
 
     public function __construct(FileLocatorInterface $fileLocator, bool $keepFirstTag = false)
     {
-        $this->keepFirstTag = $keepFirstTag;
+        $this->converter = new Converter(['preserveFirstTag' => $keepFirstTag]);
         parent::__construct($fileLocator);
     }
 
@@ -34,23 +42,19 @@ class XmlFileLoader extends FileLoader
      *
      * @param mixed $resource The resource to load.
      * @param string|null $type The resource type.
-     * @return array
+     * @return array<int|string,mixed>
      * @throws ConverterException If an error occurs while parsing the xml.
      * @throws ConfigurationBuilderException If an error occurs while reading the xml file.
-     *
-     * @psalm-suppress PossiblyInvalidArgument FileLocator::locate() returns string, since 3rd argument isn't false
      */
     public function load(mixed $resource, ?string $type = null): array
     {
-        $xmlContent = file_get_contents($this->getLocator()->locate($resource));
+        $xmlContent = @file_get_contents($this->getLocator()->locate($resource));
 
-        if ($xmlContent === '') {
-            return [];
-        }
-
-        $converter = new Converter(['preserveFirstTag' => $this->keepFirstTag]);
-
-        return $converter->convert($xmlContent);
+        return match ($xmlContent) {
+            false => throw new ConfigurationBuilderException("The configuration file '$resource' is not readable."),
+            '' => [],
+            default => $this->converter->convert($xmlContent),
+        };
     }
 
     /**
@@ -62,6 +66,6 @@ class XmlFileLoader extends FileLoader
      */
     public function supports(mixed $resource, ?string $type = null): bool
     {
-        return str_ends_with((string)$resource, '.xml') || str_ends_with((string)$resource, '.xml.dist');
+        return str_ends_with((string) $resource, '.xml') || str_ends_with((string) $resource, '.xml.dist');
     }
 }