+++ tests/Unit/Loader/PhpFileLoaderTest.php
@@ -1,72 +1,99 @@
-<?php declare(strict_types=1);
+<?php
+
+declare(strict_types=1);
+
 /*
- * Apache-2 License.
- * This file is part of susina/config-builder package, release under the APACHE-2 license.
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
+ * Copyright (c) Cristiano Cinotti
+ *
+ * This file is part of susina/config-builder package, released under the APACHE-2 license.
+ * For the full copyright and license information, please view the LICENSE file distributed
+ * with this source code.
  */
+
+namespace Susina\ConfigBuilder\Tests\Unit\Loader;
 
 use org\bovigo\vfs\vfsStream;
 use Susina\ConfigBuilder\Exception\ConfigurationBuilderException;
-use Susina\ConfigBuilder\FileLocator;
 use Susina\ConfigBuilder\Loader\PhpFileLoader;
+use Susina\ConfigBuilder\Tests\TestCase;
 use Symfony\Component\Config\Exception\FileLocatorFileNotFoundException;
+use Symfony\Component\Config\FileLocator;
 
-beforeEach(function () {
-    $this->loader = new PhpFileLoader(new FileLocator($this->getRoot()->url()));
-});
+class PhpFileLoaderTest extends TestCase
+{
+    private $loader;
 
-test('Supported php extensions', function () {
-    expect($this->loader->supports('foo.php'))->toBeTrue()
-        ->and($this->loader->supports('foo.php.dist'))->toBeTrue()
-        ->and($this->loader->supports('foo.foo'))->toBeFalse()
-        ->and($this->loader->supports('foo.foo.dist'))->toBeFalse()
-    ;
-});
+    public function setUp(): void
+    {
+        $this->loader = new PhpFileLoader(new FileLocator($this->root->url()));
+    }
 
-test('Load php file', function () {
-    $content = <<<EOF
-<?php
+    public function testSupportedExtensions(): void
+    {
+        $this->assertTrue($this->loader->supports("foo.php"));
+        $this->assertTrue($this->loader->supports("foo.php.dist"));
+        $this->assertFalse($this->loader->supports('foo.bar'));
+        $this->assertFalse($this->loader->supports('foo.bar.dist'));
+    }
 
-    return array('foo' => 'bar', 'bar' => 'baz');
+    public function testLoadPhpFile(): void
+    {
+        $expected = [
+            'foo' => 'bar',
+            'bar' => 'baz',
+        ];
 
-EOF;
-    vfsStream::newFile('parameters.php')->at($this->getRoot())->setContent($content);
-    $test = $this->loader->load('parameters.php');
+        $content = "<?php return ['foo' => 'bar', 'bar' => 'baz'];";
 
-    expect($test['foo'])->toBe('bar')
-        ->and($test['bar'])->toBe('baz');
-});
+        vfsStream::newFile('parameters.php')->at($this->root)->setContent($content);
+        $actual = $this->loader->load('parameters.php');
 
-test('Load not existent php file', function () {
-    $this->loader->load('inexistent.php');
-})->throws(FileLocatorFileNotFoundException::class, 'The file "inexistent.php" does not exist (in: "vfs://root").');
+        $this->assertSame($expected, $actual);
+    }
 
-test('Load php file with invalid conten', function () {
-    $content = <<<EOF
+    public function testLoadNotExistentPhpFileThrowsException(): void
+    {
+        $this->expectException(FileLocatorFileNotFoundException::class);
+        $this->expectExceptionMessageIs('The file "inexistent.php" does not exist (in: "vfs://root").');
+
+        $this->loader->load('inexistent.php');
+    }
+
+    public function testLoadPhpFileWithInvalidContentThrowsException(): void
+    {
+        $this->expectException(ConfigurationBuilderException::class);
+        $this->expectExceptionMessageIsOrContains("The configuration file 'nonvalid.php' has invalid content.");
+
+        $content = <<<EOF
 not php content
 only plain
 text
 EOF;
-    vfsStream::newFile('nonvalid.php')->at($this->getRoot())->setContent($content);
-    $this->loader->load('nonvalid.php');
-})->throws(ConfigurationBuilderException::class, "The configuration file 'nonvalid.php' has invalid content.");
+        vfsStream::newFile('nonvalid.php')->at($this->root)->setContent($content);
+        $this->loader->load('nonvalid.php');
+    }
 
-test('Load empty php file', function () {
-    vfsStream::newFile('empty.php')->at($this->getRoot())->setContent('');
-    $actual = $this->loader->load('empty.php');
+    public function testLoadEmptyPhpFileReturnsEmptyArray(): void
+    {
+        vfsStream::newFile('empty.php')->at($this->root)->setContent('');
+        $actual = $this->loader->load('empty.php');
 
-    expect($actual)->toBe([]);
-});
+        $this->assertIsArray($actual);
+        $this->assertEmpty($actual);
+    }
 
-test('Load not readable php file', function () {
-    $content = <<<EOF
-<?php
+    public function testLoadNotReadablePhpFileThrowsException(): void
+    {
+        if ($this->runnungOnWindows()) {
+            $this->markTestSkipped("Changing file permission doesn't work on Windows");
+        }
 
-    return array('foo' => 'bar', 'bar' => 'baz');
+        $this->expectException(ConfigurationBuilderException::class);
+        $this->expectExceptionMessageIs("The configuration file 'notreadable.php' does not exist or is not readable.");
 
-EOF;
-    vfsStream::newFile('notreadable.php', 200)->at($this->getRoot())->setContent($content);
-    $this->loader->load('notreadable.php');
-})->throws(ConfigurationBuilderException::class, 'Path "vfs://root/notreadable.php" was expected to be readable.')
-    ->skipOnWindows();
+        $content = "<?php return array('foo' => 'bar', 'bar' => 'baz');";
+        vfsStream::newFile('notreadable.php', 200)->at($this->root)->setContent($content);
+
+        $actual = $this->loader->load('notreadable.php');
+    }
+}